# doc/api-docs-smart/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PinShop Server",
    "version": "v1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
...
  "paths": {
    "/api/v1/user/login": {
      "post": {
        "summary": "User login\nSupports WeChat openid style login",
        "deprecated": false,
        "tags": [
          "UserController"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/51cb5a19fbeb4ccda59cbfab5e696243"
              },
              "examples": {
                "json": {
                  "summary": "test data",
                  "value": "{\n  \"openid\": \"\",\n  \"nickname\": \"\",\n  \"avatarUrl\": \"\"\n}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/d4a93d5b69e349f7884016b0f02b403d"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"code\": 200,\n  \"message\": \"success\",\n  \"data\": 0,\n  \"timestamp\": 1709823456789\n}"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/da0c932b0e6d3b90859201037a7929ee"
                },
                "examples": {
                  "json": {
                    "summary": "response example",
                    "value": "{\n  \"code\": 200,\n  \"message\": \"success\",\n  \"data\": {},\n  \"timestamp\": 1709823456789\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "api-v1-user-login-POST",
        "parameters": [
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "header",
            "description": "认证Token (Bearer {jwt_token})",
            "required": false,
            "name": "Authorization"
          }
        ]
      }
    },
    "/api/v1/user/profile": {
      "put": {
        "summary": "Update user profile",
        "deprecated": false,
        "tags": [
          "UserController"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/3c7d48c32d0c7cd9413b0757567073b8"
              },
              "examples": {
                "json": {
                  "summary": "test data",
                  "value": "{\n  \"userId\": 0,\n  \"nickname\": \"\",\n  \"phone\": \"\"\n}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/2e2e10f941a0ac19f25c4a7520d458b1"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"code\": 200,\n  \"message\": \"success\",\n  \"timestamp\": 1709823456789\n}"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/da0c932b0e6d3b90859201037a7929ee"
                },
                "examples": {
                  "json": {
                    "summary": "response example",
                    "value": "{\n  \"code\": 200,\n  \"message\": \"success\",\n  \"data\": {},\n  \"timestamp\": 1709823456789\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "api-v1-user-profile-PUT",
        "parameters": [
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "header",
            "description": "认证Token (Bearer {jwt_token})",
            "required": false,
            "name": "Authorization"
          }
        ]
      },
      "get": {
        "summary": "Get user profile",
        "deprecated": false,
        "tags": [
          "UserController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/cd156cb8b87124404c84c30e3d09e7d6"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"code\": 200,\n  \"message\": \"success\",\n  \"data\": {\n    \"userId\": 0,\n    \"nickname\": \"\",\n    \"avatarUrl\": \"\",\n    \"phone\": \"\",\n    \"status\": 0\n  },\n  \"timestamp\": 1709823456789\n}"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/da0c932b0e6d3b90859201037a7929ee"
                },
                "examples": {
                  "json": {
                    "summary": "response example",
                    "value": "{\n  \"code\": 200,\n  \"message\": \"success\",\n  \"data\": {},\n  \"timestamp\": 1709823456789\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "api-v1-user-profile-GET",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "number"
            },
            "in": "query",
            "description": "user ID",
            "required": true,
            "example": "0",
            "name": "userId"
          },
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "header",
            "description": "认证Token (Bearer {jwt_token})",
            "required": false,
            "name": "Authorization"
          }
        ]
      }
    },
    "/api/v1/user/addresses": {
      "get": {
        "summary": "Get user address list",
        "deprecated": false,
        "tags": [
          "UserController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/f84fd0322537c36819a01436a0dbd3b6"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"code\": 200,\n  \"message\": \"success\",\n  \"data\": [\n    {\n      \"addressId\": 0,\n      \"receiverName\": \"\",\n      \"receiverPhone\": \"\",\n      \"province\": \"\",\n      \"city\": \"\",\n      \"district\": \"\",\n      \"detailAddr\": \"\",\n      \"isDefault\": 0\n    }\n  ],\n  \"timestamp\": 1709823456789\n}"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/da0c932b0e6d3b90859201037a7929ee"
                },
                "examples": {
                  "json": {
                    "summary": "response example",
                    "value": "{\n  \"code\": 200,\n  \"message\": \"success\",\n  \"data\": {},\n  \"timestamp\": 1709823456789\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "api-v1-user-addresses-GET",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "number"
            },
            "in": "query",
            "description": "user ID",
            "required": true,
            "example": "0",
            "name": "userId"
          },
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "header",
            "description": "认证Token (Bearer {jwt_token})",
            "required": false,
            "name": "Authorization"
          }
        ]
      }
    },
    "/api/v1/user/address": {
      "post": {
        "summary": "Add user address",
        "deprecated": false,
        "tags": [
          "UserController"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/a4cabcfc8043619a3d7b6730f7209b54"
              },
              "examples": {
                "json": {
                  "summary": "test data",
                  "value": "{\n  \"userId\": 0,\n  \"receiverName\": \"\",\n  \"receiverPhone\": \"\",\n  \"province\": \"\",\n  \"city\": \"\",\n  \"district\": \"\",\n  \"detailAddr\": \"\",\n  \"isDefault\": 0\n}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/d4a93d5b69e349f7884016b0f02b403d"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"code\": 200,\n  \"message\": \"success\",\n  \"data\": 0,\n  \"timestamp\": 1709823456789\n}"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/da0c932b0e6d3b90859201037a7929ee"
                },
                "examples": {
                  "json": {
                    "summary": "response example",
                    "value": "{\n  \"code\": 200,\n  \"message\": \"success\",\n  \"data\": {},\n  \"timestamp\": 1709823456789\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "api-v1-user-address-POST",
        "parameters": [
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "header",
            "description": "认证Token (Bearer {jwt_token})",
            "required": false,
            "name": "Authorization"
          }
        ]
      }
    },
    "/api/v1/user/address/{addressId}": {
      "delete": {
        "summary": "Delete user address",
        "deprecated": false,
        "tags": [
          "UserController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/2e2e10f941a0ac19f25c4a7520d458b1"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"code\": 200,\n  \"message\": \"success\",\n  \"timestamp\": 1709823456789\n}"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/da0c932b0e6d3b90859201037a7929ee"
                },
                "examples": {
                  "json": {
                    "summary": "response example",
                    "value": "{\n  \"code\": 200,\n  \"message\": \"success\",\n  \"data\": {},\n  \"timestamp\": 1709823456789\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "api-v1-user-address-{addressId}-DELETE",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "number"
            },
            "in": "path",
            "description": "address ID",
            "required": true,
            "example": "0",
            "name": "addressId"
          },
          {
            "schema": {
              "format": "int64",
              "type": "number"
            },
            "in": "query",
            "description": "user ID",
            "required": true,
            "example": "0",
            "name": "userId"
          },
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "header",
            "description": "认证Token (Bearer {jwt_token})",
            "required": false,
            "name": "Authorization"
          }
        ]
      },
      "put": {
        "summary": "Update user address",
        "deprecated": false,
        "tags": [
          "UserController"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/052b58e5a33e0d7dea116c6e72c53376"
              },
              "examples": {
                "json": {
                  "summary": "test data",
                  "value": "{\n  \"userId\": 0,\n  \"receiverName\": \"\",\n  \"receiverPhone\": \"\",\n  \"province\": \"\",\n  \"city\": \"\",\n  \"district\": \"\",\n  \"detailAddr\": \"\",\n  \"isDefault\": 0\n}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/2e2e10f941a0ac19f25c4a7520d458b1"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"code\": 200,\n  \"message\": \"success\",\n  \"timestamp\": 1709823456789\n}"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/da0c932b0e6d3b90859201037a7929ee"
                },
                "examples": {
                  "json": {
                    "summary": "response example",
                    "value": "{\n  \"code\": 200,\n  \"message\": \"success\",\n  \"data\": {},\n  \"timestamp\": 1709823456789\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "api-v1-user-address-{addressId}-PUT",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "number"
            },
            "in": "path",
            "description": "address ID",
            "required": true,
            "example": "0",
            "name": "addressId"
          },
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "header",
            "description": "认证Token (Bearer {jwt_token})",
            "required": false,
            "name": "Authorization"
          }
        ]
      }
    },
    "/api/v1/user/address/{addressId}/default": {
      "post": {
        "summary": "Set default address",
        "deprecated": false,
        "tags": [
          "UserController"
        ],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "format": "int64",
                    "description": "user ID",
                    "type": "number",
                    "example": "0"
                  }
                },
                "required": [
                  "userId"
                ]
              },
              "examples": {
                "json": {
                  "summary": "test data",
                  "value": "&0"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/2e2e10f941a0ac19f25c4a7520d458b1"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"code\": 200,\n  \"message\": \"success\",\n  \"timestamp\": 1709823456789\n}"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/da0c932b0e6d3b90859201037a7929ee"
                },
                "examples": {
                  "json": {
                    "summary": "response example",
                    "value": "{\n  \"code\": 200,\n  \"message\": \"success\",\n  \"data\": {},\n  \"timestamp\": 1709823456789\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "api-v1-user-address-{addressId}-default-POST",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "number"
            },
            "in": "path",
            "description": "address ID",
            "required": true,
            "example": "0",
            "name": "addressId"
          },
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "header",
            "description": "认证Token (Bearer {jwt_token})",
            "required": false,
            "name": "Authorization"
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "dfcaca810592df45d261b3577fc9fc3f": {
        "properties": {
          "userId": {
            "format": "int64",
            "description": "No comments found.",
            "type": "number",
            "example": "0"
          },
          "nickname": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "avatarUrl": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "phone": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "status": {
            "format": "int32",
            "description": "No comments found.",
            "type": "integer",
            "example": "0"
          }
        }
      },
...
      "cd156cb8b87124404c84c30e3d09e7d6": {
        "properties": {
          "code": {
            "format": "int32",
            "description": "业务状态码",
            "type": "integer",
            "example": "200"
          },
          "message": {
            "format": "string",
            "description": "响应消息",
            "type": "string",
            "example": "success"
          },
          "data": {
            "description": "业务数据(object)",
            "$ref": "#/components/schemas/dfcaca810592df45d261b3577fc9fc3f"
          },
          "timestamp": {
            "format": "int64",
            "description": "响应时间戳(毫秒)",
            "type": "number",
            "example": "1709823456789"
          }
        }
      },
...
      "2e2e10f941a0ac19f25c4a7520d458b1": {
        "properties": {
          "code": {
            "format": "int32",
            "description": "业务状态码",
            "type": "integer",
            "example": "200"
          },
          "message": {
            "format": "string",
            "description": "响应消息",
            "type": "string",
            "example": "success"
          },
          "data": {
            "description": "业务数据(object)",
            "type": "object"
          },
          "timestamp": {
            "format": "int64",
            "description": "响应时间戳(毫秒)",
            "type": "number",
            "example": "1709823456789"
          }
        }
      },
...
      "96e8a5750715bcc8925cbe172b1a8093": {
        "properties": {
          "addressId": {
            "format": "int64",
            "description": "No comments found.",
            "type": "number",
            "example": "0"
          },
          "receiverName": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "receiverPhone": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "province": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "city": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "district": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "detailAddr": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "isDefault": {
            "format": "int32",
            "description": "No comments found.",
            "type": "integer",
            "example": "0"
          }
        }
      },
      "f84fd0322537c36819a01436a0dbd3b6": {
        "properties": {
          "code": {
            "format": "int32",
            "description": "业务状态码",
            "type": "integer",
            "example": "200"
          },
          "message": {
            "format": "string",
            "description": "响应消息",
            "type": "string",
            "example": "success"
          },
          "data": {
            "description": "业务数据",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/96e8a5750715bcc8925cbe172b1a8093"
            }
          },
          "timestamp": {
            "format": "int64",
            "description": "响应时间戳(毫秒)",
            "type": "number",
            "example": "1709823456789"
          }
        }
      },
...
      "d4a93d5b69e349f7884016b0f02b403d": {
        "properties": {
          "code": {
            "format": "int32",
            "description": "业务状态码",
            "type": "integer",
            "example": "200"
          },
          "message": {
            "format": "string",
            "description": "响应消息",
            "type": "string",
            "example": "success"
          },
          "data": {
            "format": "int64",
            "description": "业务数据",
            "type": "number"
          },
          "timestamp": {
            "format": "int64",
            "description": "响应时间戳(毫秒)",
            "type": "number",
            "example": "1709823456789"
          }
        }
      },
...
      "052b58e5a33e0d7dea116c6e72c53376": {
        "properties": {
          "userId": {
            "format": "int64",
            "description": "No comments found.",
            "type": "number",
            "example": "0"
          },
          "receiverName": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "receiverPhone": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "province": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "city": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "district": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "detailAddr": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "isDefault": {
            "format": "int32",
            "description": "No comments found.",
            "type": "integer",
            "example": "0"
          }
        }
      },
...
      "51cb5a19fbeb4ccda59cbfab5e696243": {
        "properties": {
          "openid": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "nickname": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "avatarUrl": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          }
        }
      },
      "3c7d48c32d0c7cd9413b0757567073b8": {
        "properties": {
          "userId": {
            "format": "int64",
            "description": "No comments found.",
            "type": "number",
            "example": "0"
          },
          "nickname": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "phone": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          }
        }
      },
...
      "da0c932b0e6d3b90859201037a7929ee": {
        "properties": {
          "code": {
            "format": "int32",
            "description": "业务状态码",
            "type": "integer",
            "example": "200"
          },
          "message": {
            "format": "string",
            "description": "响应消息",
            "type": "string",
            "example": "success"
          },
          "data": {
            "description": "业务数据(object)",
            "type": "object"
          },
          "timestamp": {
            "format": "int64",
            "description": "响应时间戳(毫秒)",
            "type": "number",
            "example": "1709823456789"
          }
        }
      },
      "a4cabcfc8043619a3d7b6730f7209b54": {
        "properties": {
          "userId": {
            "format": "int64",
            "description": "No comments found.",
            "type": "number",
            "example": "0"
          },
          "receiverName": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "receiverPhone": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "province": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "city": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "district": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "detailAddr": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "isDefault": {
            "format": "int32",
            "description": "No comments found.",
            "type": "integer",
            "example": "0"
          }
        }
      }
    }
  }
}